pyproject.toml
[project]
name = "mcp-prompt-tester"
version = "0.1.0"
description = "MCP server for testing LLM prompts with different providers"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Prompt Tester" }]
dependencies = [
    "anyio>=4.5",
    "click>=8.1.0",
    "mcp>=1.3.0",
    "openai>=1.64.0",
    "anthropic>=0.47.2",
    "python-dotenv>=1.0.0"
]

[project.scripts]
prompt-tester = "mcp_prompt_tester.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcp_prompt_tester"]

[tool.pyright]
include = ["mcp_prompt_tester"]
venvPath = "."
venv = ".venv"

[tool.ruff]
line-length = 100
target-version = "py310" 